refactor(backend): colocate fetch URL allowlist in web_tools to avoid package growth
