+++ backend/utils/retrieval/agentic.py
@@ -50,7 +50,7 @@ from utils.retrieval.tools import (
 )
 from utils.retrieval.tools.app_tools import load_app_tools, get_tool_status_message
 from utils.retrieval.tool_result_boundaries import preserve_chat_memory_tool_result_boundary
-from utils.retrieval.fetch_url_allowlist import (
+from utils.retrieval.tools.web_tools import (
     extract_user_turn_urls,
     user_url_allowlist_block,
 )
